Refactor primary Build routine
The multi-driver plumbing in the Build routine was making this routine fragile
and difficult to adapt to native multi-arch support for the kubernetes driver.
This doesn't completely remove the Driver abstraction, but should help for
maintainability.  This does not implement native multi-arch yet (more changes
will be required)

I also added a bit more coverage to make sure I didn't cause any regressions.